+++ src/packages/block/modals/manifest-viewer/manifest.ts
@@ -1,4 +1,4 @@
-import type { ManifestModal } from '@umbraco-cms/backoffice/extension-registry';
+import type { ManifestModal } from '@umbraco-cms/backoffice/modal';
 
 export const manifest: ManifestModal = {
 	type: 'modal',
