+++ src/packages/core/extension-registry/interfaces/index.ts
@@ -2,7 +2,6 @@ export * from './dashboard-element.interface.js';
 export * from './external-login-provider-element.interface.js';
 export * from './file-upload-preview.interface.js';
 export * from './menu-item-element.interface.js';
-export * from './modal-extension-element.interface.js';
 export * from './property-editor-ui-element.interface.js';
 export * from './section-element.interface.js';
 export * from './section-sidebar-app-element.interface.js';
