+++ src/packages/core/modal/component/modal-base.element.ts
@@ -1,7 +1,7 @@
 import type { UmbModalRejectReason, UmbModalContext } from '../context/index.js';
 import { property } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
-import type { ManifestModal, UmbModalExtensionElement } from '@umbraco-cms/backoffice/extension-registry';
+import type { ManifestModal, UmbModalExtensionElement } from '@umbraco-cms/backoffice/modal';
 
 export abstract class UmbModalBaseElement<
 		ModalDataType extends object = object,
