+++ src/packages/core/modal/extensions/index.ts
@@ -1,0 +1,9 @@
+export type * from './modal-extension-element.interface.js';
+export type * from './modal.extension.js';
+import type { ManifestModal } from './modal.extension.js';
+
+declare global {
+	interface UmbExtensionManifestMap {
+		UmbModalExtension: ManifestModal;
+	}
+}
