+++ src/packages/core/tree/entity-actions/duplicate-to/modal/manifests.ts
@@ -1,5 +1,5 @@
 import { UMB_DUPLICATE_TO_MODAL_ALIAS } from './constants.js';
-import type { ManifestModal } from '@umbraco-cms/backoffice/extension-registry';
+import type { ManifestModal } from '@umbraco-cms/backoffice/modal';
 
 export const manifests: Array<ManifestModal> = [
 	{
